% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_comorbs.R
\name{extract_comorbs}
\alias{extract_comorbs}
\title{Register-based CoMorbidities in a SWEdish setting (rcmswe)}
\usage{
extract_comorbs(
  search_df,
  sqlite_path,
  sqlite_NPR_name = "PAR",
  sqlite_LMED_name = "LMED",
  NPR = TRUE,
  LMED = FALSE,
  CCI = TRUE
)
}
\arguments{
\item{search_df}{A data-frame with two columns, column one should be study ID's and column two should be an index date.}

\item{sqlite_path}{A character vector specifying the path to an SQLite3 database containing registers.}

\item{sqlite_NPR_name}{A character vector specifying the name of the NPR inside the SQLite database (currently not functioning)}

\item{sqlite_LMED_name}{A character vector specifying the name of the Prescribed Drug Register inside the SQLite database (currently not functioning)}

\item{NPR}{A logical parameter dictating if the returning dataset should contain individual groups of co-morbidities (default = TRUE)}

\item{LMED}{A logical parameter dictating if the returning dataset should be expanded with data from LMED (default = FALSE, currently non-functioning)}

\item{CCI}{A logical parameter dictating if the returning dataset should contain columns for weighted and unweighted CCI (currently solely based on NPR data)}
}
\description{
Extract Co-Morbidities from Swedish Population Health Registers
}
